Use multistage build
There is some oddity in how node 9.8 is running on modern debian when it comes to `armv7l`. `npm install` gets `Unknown system error -75` which apparently is some overflow issue. Since building works in the `node:9.8` image, we should be able to copy the built `node_modules` to the `base` image to run with the appropriate ALSA libraries.